Fix brightness udev rule
$DISPLAY is not set in udev's environment, so we'll have to set it manually. On my laptop, it is
always :0, so depending on one value (while hackish) works for me. The same goes for the username: X
needs to know where to find an .Xauthority file and on my laptop, only one user and X session is
active at the same time.

xset will still exit with a failure when this udev rule is triggered upon boot.